tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/062.html

<!DOCTYPE html>
<html><head>
  <title> scheduler: cloneNode (shallow) and script execution</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
  <script>log('HEAD script');</script>
</head>
<body>

  <div id="log">FAILED (This TC requires JavaScript enabled)</div>

  <script>log('inline script #1');
    var script=document.getElementsByTagName('script')[3];
    document.body.appendChild( script.cloneNode(false) ).appendChild(document.createTextNode('log("clone");'));
    log('end script #1');
  </script>
  <script type="text/javascript">
  log( 'inline script #2' );
  var t = async_test()

  function test() {
    assert_array_equals(eventOrder, ['HEAD script', 'inline script #1', 'end script #1', 'inline script #2']);
    t.done();
  }
  onload = t.step_func(test);
  </script>

</body></html>
